DEVX-2203: replace -CAcreateserial with parallel-safe random (#267)
DEVX-2203: replace -CAcreateserial with parallel-safe random -set_serial.

openssl x509 ... option -CAcreateserial isn't parallel-friendly,  and cert-generation
was recently parallelized.  This created an overwrite/corruption condition on the
file used by -CAcreateserial, in our case:

>  unable to load number from ./snakeoil-ca-1.srl

Whilst a serial can be loaded from a specific file via -CAserial [file], it's
not clear how to interact this with -CAcreateserial option for a unique filename,
if that's even possible.

The serial has to be unique and <= 20 octets.  openssl seems to be generating
monotonic-increasing (time-based?) ones, but 8 octests random is plenty for our
demo certs.

Replaced with an awk script generating 8 octets, pass to -set_serial for each
cert.

Co-authored-by: Yeva Byzek <ybyzek@users.noreply.github.com>