DEVX-2720: fixes for failed runs on modern docker and docker-compose
I have not been able to isolate why these changes are required on my development machine. However,
without these the cp-demo start script failed for a variety of reasons.  The return values from
executions of docker-compose exec ... resulted in empty variables which evaluated to various
runtime errors in the start script.

This set of changes does not directly address the issues identified in DEVX-2720 but instead was required
by me to get the demo to run at all.  I will next attempt to understand and address the issues noted in
the Jira issue.